requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
meilisearch-python-sdk>=4.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
anyio>=3.7.1
